+++ packages/components/addon/components/hds/dropdown/toggle-icon.js
@@ -1,0 +1,59 @@
+import Component from '@glimmer/component';
+import { assert } from '@ember/debug';
+
+export default class HdsDropdownToggleIconComponent extends Component {
+  /**
+   * @param text
+   * @type {string}
+   * @description The text of the `aria-label` applied to the toggle
+   */
+  get text() {
+    let { text } = this.args;
+
+    assert(
+      '@text for "Hds::Dropdown::ToggleIcon" must have a valid value',
+      text !== undefined
+    );
+
+    return text;
+  }
+
+  /**
+   * Indicates if a dropdown chevron icon should be displayed; should be displayed unless the "more-horizontal" icon is used.
+   *
+   * @param hasChevron
+   * @type {boolean}
+   * @default true
+   */
+  get hasChevron() {
+    return this.args.hasChevron ?? true;
+  }
+
+  /**
+   * Sets the icon name
+   *
+   * @param icon
+   * @type {string}
+   * @default user
+   */
+  get icon() {
+    return this.args.icon ?? 'user';
+  }
+
+  /**
+   * @param onClick
+   * @type {function}
+   * @default () => {}
+   */
+  get onClick() {
+    let { onClick } = this.args;
+
+    // notice: this is a guard used in case the toggle is used as standalone element (eg. in the showcase)
+    // in reality it's always used inside the Dropdown main component as yielded component, so the onClick handler is always defined
+    if (typeof onClick === 'function') {
+      return onClick;
+    } else {
+      return () => {};
+    }
+  }
+}