+++ packages/components/tests/dummy/app/router.js
@@ -22,11 +22,14 @@ Router.map(function () {
     this.route('card');
     this.route('icon-tile');
     this.route('link', function () {
-      this.route('standalone', { path: '/' });
+      this.route('standalone');
+      this.route('cta');
     });
     this.route('link-to', function () {
-      this.route('standalone', { path: '/' });
+      this.route('standalone');
+      this.route('cta');
     });
+    this.route('dropdown');
   });
   this.route('content', function () {
     this.route('writing-guidelines');